+++ src/main/java/com/pede/ai/infra/inbounds/forms/ProductForm.java
@@ -3,12 +3,13 @@ package com.pede.ai.infra.inbounds.forms;
 import com.pede.ai.core.commons.enums.ProductType;
 import com.pede.ai.core.domain.product.DomainProduct;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record ProductForm(
     String name,
     String desc,
-    float price,
+    BigDecimal price,
     ProductType type
 ) {
